% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bellman_ford.R
\name{BellmanFord}
\alias{BellmanFord}
\title{BellmanFord naif Algorithm to find shortest path from a node to others}
\usage{
BellmanFord(vertices, edges, start)
}
\arguments{
\item{vertices}{vector contains names of each node. Names can be a int or str.}

\item{edges}{Data.frame object with three variables (from, to and weight)
that contains the edges of the graph_DataFrame (from, to) with the weight of the edge (w)}

\item{start}{initial node (could be int or str).}
}
\value{
list of two elements.
Distance: the shortest distance from the start node to other nodes (vector)
parents : the predecessor of each node (vector)
}
\description{
BellmanFord naif Algorithm to find shortest path from a node to others
}
\examples{
vertices <- c("r", "a", "b", "c")
edges <- data.frame(from   = c("r", "a", "b", "c", "c"),
                    to = c("a", "b", "c", "r", "a"),
                    weight = c(3, 1, 2, 7, -4))
start="r"
}
